chore: move AddWhereCondModal to legacy
